backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.4
mysql-connector-python==8.3.0
beautifulsoup4==4.12.3
requests==2.31.0
torch==2.2.1+cpu
transformers==4.38.2
